cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gpunet-cli"
version = "0.1.0"
description = "Command-line interface for the GPU Network platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.scripts]
gpunet = "gpunet_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["gpunet_cli"]
include = ["skills/**/*"]
